Refactoring CartDataProiver
