+++ Services/WebStore.Interfaces/DataProviders/ICartStore.cs
@@ -1,0 +1,9 @@
+﻿using WebStore.Domain.Entity;
+
+namespace WebStore.Interfaces.DataProviders
+{
+    public interface ICartStore
+    {
+        Cart Cart { get; set; } 
+    }
+}
