+++ Services/WebStore.Services/DataProviders/CookiesDataProvider/CartStore.cs
@@ -1,0 +1,50 @@
+﻿using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using WebStore.Domain.Entity;
+using WebStore.Interfaces.DataProviders;
+
+namespace WebStore.Services.DataProviders.CookiesDataProvider
+{
+    public class CartStore : ICartStore
+    {
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        private readonly string _cartName;
+
+        public Cart Cart
+        {
+            get
+            {
+                var context = _httpContextAccessor.HttpContext;
+                var cookies = context.Response.Cookies;
+                var cart_cookie = context.Request.Cookies[_cartName];
+                if (cart_cookie is null)
+                {
+                    var cart = new Cart();
+                    cookies.Append(_cartName, JsonConvert.SerializeObject(cart));
+                    return cart;
+                }
+
+                ReplaceCookie(cookies, cart_cookie);
+                return JsonConvert.DeserializeObject<Cart>(cart_cookie);
+            }
+            set => ReplaceCookie(_httpContextAccessor.HttpContext.Response.Cookies, JsonConvert.SerializeObject(value));
+        }
+
+        public CartStore(IHttpContextAccessor httpContextAccessor)
+        {
+            _httpContextAccessor = httpContextAccessor;
+            var user = _httpContextAccessor.HttpContext.User;
+            var user_name = user.Identity.IsAuthenticated ? user.Identity.Name : null;
+            _cartName = $"cart[{user_name}]";
+        }
+
+        private void ReplaceCookie(IResponseCookies cookies, string cookie)
+        {
+            cookies.Delete(_cartName);
+            cookies.Append(_cartName, cookie, new CookieOptions { Expires = DateTime.Now.AddDays(15) });
+        }
+    }
+}
