+++ Services/WebStore.Services/DataProviders/CookiesDataProvider/CookieCartProvider.cs
@@ -1,7 +1,4 @@
-﻿using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using WebStore.Domain.Entity;
 using WebStore.Interfaces.Api;
@@ -12,82 +9,53 @@ namespace WebStore.Services.DataProviders.CookiesDataProvider
     public class CookieCartProvider : ICartDataProvider
     {
         private readonly IProductService _productDataProvider;
-        private readonly IHttpContextAccessor _httpContextAccessor;
-        private readonly string _cartName;
+        private readonly ICartStore _cartStore;        
 
-        public Cart Cart 
-        {
-            get 
-            {
-                var context = _httpContextAccessor.HttpContext;
-                var cookies = context.Response.Cookies;
-                var cart_cookie = context.Request.Cookies[_cartName];
-                if (cart_cookie is null)
-                {
-                    var cart = new Cart();
-                    cookies.Append(_cartName, JsonConvert.SerializeObject(cart));
-                    return cart;
-                }
-
-                ReplaceCookie(cookies, cart_cookie);
-                return JsonConvert.DeserializeObject<Cart>(cart_cookie);
-            }
-            private set => ReplaceCookie(_httpContextAccessor.HttpContext.Response.Cookies, JsonConvert.SerializeObject(value));
-        }
-        private void ReplaceCookie(IResponseCookies cookies, string cookie)
-        {
-            cookies.Delete(_cartName);
-            cookies.Append(_cartName, cookie, new CookieOptions { Expires = DateTime.Now.AddDays(15) });
-        }
-
-        public CookieCartProvider(IProductService productDataProvider,IHttpContextAccessor httpContextAccessor)
+        public CookieCartProvider(IProductService productDataProvider,ICartStore cartStore)
         {
             _productDataProvider = productDataProvider;
-            _httpContextAccessor = httpContextAccessor;
-            var user = _httpContextAccessor.HttpContext.User;
-            var user_name = user.Identity.IsAuthenticated ? user.Identity.Name : null;
-            _cartName = $"cart[{user_name}]";
+            _cartStore = cartStore;
         }
         public void AddToCart(int productId)
         {
-            var cart = Cart;
+            var cart = _cartStore.Cart;
             var product=cart.Items?.FirstOrDefault(p => p.ProductId == productId);
             if (product is null)
                 cart.Items.Add(new CartItem { ProductId = productId, Quantity = 1 });
             else
                 product.Quantity++;
-            Cart = cart;
+            _cartStore.Cart = cart;
         }
 
         public void ClearCart()
         {
-            var cart = Cart;
+            var cart = _cartStore.Cart;
             cart.Items.Clear();
-            Cart = cart;
+            _cartStore.Cart = cart;
         }
 
         public void DecreaseFromCart(int productId)
         {
-            var cart = Cart;
+            var cart = _cartStore.Cart;
             var product = cart.Items.FirstOrDefault(p => p.ProductId == productId);
             if (product is null) return;
             if (product.Quantity > 1) product.Quantity--;
             if (product.Quantity < 2) cart.Items.Remove(product);
-            Cart = cart;
+            _cartStore.Cart = cart;
         }
 
         public void DeleteFromCart(int productId)
         {
-            var cart = Cart;
+            var cart = _cartStore.Cart;
             var product = cart.Items.FirstOrDefault(p => p.ProductId == productId);
             if (product is null) return;
             cart.Items.Remove(product);
-            Cart = cart;
+            _cartStore.Cart = cart;
         }
 
         public IEnumerable<CartItem> GetCartItems()
         {
-            var cart = Cart;
+            var cart = _cartStore.Cart;
             return cart.Items;
         }
     }
